refactoring the routes to use the newly refactored Models
