+++ src/device-registry/routes/v1/events.js
@@ -6,18 +6,7 @@ const constants = require("@config/constants");
 const mongoose = require("mongoose");
 const { logElement, logText, logObject } = require("@utils/log");
 const ObjectId = mongoose.Types.ObjectId;
-const { getModelByTenant } = require("@config/database");
-
-const NetworkSchema = require("@models/Network");
-const NetworkModel = (tenant) => {
-  try {
-    const networks = mongoose.model("networks");
-    return networks;
-  } catch (error) {
-    const networks = getModelByTenant(tenant, "network", NetworkSchema);
-    return networks;
-  }
-};
+const NetworkModel = require("@models/Network");
 
 const validNetworks = async () => {
   const networks = await NetworkModel("airqo").distinct("name");
