+++ src/device-registry/routes/v2/cohorts.js
@@ -7,20 +7,11 @@ const mongoose = require("mongoose");
 const ObjectId = mongoose.Types.ObjectId;
 const { logElement, logText, logObject } = require("@utils/log");
 const isEmpty = require("is-empty");
-const log4js = require("log4js");
-const logger = log4js.getLogger(`${constants.ENVIRONMENT} -- cohorts-route-v2`);
-const { getModelByTenant } = require("@config/database");
+const logger = require("log4js").getLogger(
+  `${constants.ENVIRONMENT} -- cohorts-route-v2`
+);
 
-const NetworkSchema = require("@models/Network");
-const NetworkModel = (tenant) => {
-  try {
-    const networks = mongoose.model("networks");
-    return networks;
-  } catch (error) {
-    const networks = getModelByTenant(tenant, "network", NetworkSchema);
-    return networks;
-  }
-};
+const NetworkModel = require("@models/Network");
 
 const validNetworks = async () => {
   const networks = await NetworkModel("airqo").distinct("name");
@@ -74,7 +65,7 @@ router.delete(
         .bail()
         .trim()
         .toLowerCase()
-        .custom(validateNetwork)
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -106,7 +97,7 @@ router.put(
         .bail()
         .trim()
         .toLowerCase()
-        .custom(validateNetwork)
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -158,7 +149,7 @@ router.post(
         .bail()
         .trim()
         .toLowerCase()
-        .custom(validateNetwork)
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -200,7 +191,7 @@ router.get(
       .bail()
       .trim()
       .toLowerCase()
-      .custom(validateNetwork)
+      .isIn(constants.NETWORKS)
       .withMessage("the tenant value is not among the expected ones"),
   ]),
   oneOf([
@@ -234,7 +225,7 @@ router.get(
       .bail()
       .trim()
       .toLowerCase()
-      .custom(validateNetwork)
+      .isIn(constants.NETWORKS)
       .withMessage("the tenant value is not among the expected ones"),
   ]),
   oneOf([
@@ -268,7 +259,7 @@ router.get(
       .bail()
       .trim()
       .toLowerCase()
-      .custom(validateNetwork)
+      .isIn(constants.NETWORKS)
       .withMessage("the tenant value is not among the expected ones"),
   ]),
   oneOf([
@@ -304,7 +295,7 @@ router.put(
         .bail()
         .trim()
         .toLowerCase()
-        .isIn(["kcca", "airqo"])
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -348,7 +339,7 @@ router.get(
         .bail()
         .trim()
         .toLowerCase()
-        .isIn(["kcca", "airqo"])
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -375,7 +366,7 @@ router.get(
         .bail()
         .trim()
         .toLowerCase()
-        .isIn(["kcca", "airqo"])
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -402,7 +393,7 @@ router.post(
         .bail()
         .trim()
         .toLowerCase()
-        .isIn(["kcca", "airqo"])
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -449,7 +440,7 @@ router.delete(
         .bail()
         .trim()
         .toLowerCase()
-        .isIn(["kcca", "airqo"])
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -496,7 +487,7 @@ router.delete(
         .bail()
         .trim()
         .toLowerCase()
-        .isIn(["kcca", "airqo"])
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -541,7 +532,7 @@ router.post(
         .bail()
         .trim()
         .toLowerCase()
-        .custom(validateNetwork)
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
@@ -577,7 +568,7 @@ router.get(
         .bail()
         .trim()
         .toLowerCase()
-        .custom(validateNetwork)
+        .isIn(constants.NETWORKS)
         .withMessage("the tenant value is not among the expected ones"),
     ],
   ]),
