+++ src/device-registry/routes/v2/kya.js
@@ -7,18 +7,7 @@ const mongoose = require("mongoose");
 const ObjectId = mongoose.Types.ObjectId;
 const { logElement, logText, logObject } = require("@utils/log");
 const isEmpty = require("is-empty");
-const { getModelByTenant } = require("@config/database");
-
-const NetworkSchema = require("@models/Network");
-const NetworkModel = (tenant) => {
-  try {
-    const networks = mongoose.model("networks");
-    return networks;
-  } catch (error) {
-    const networks = getModelByTenant(tenant, "network", NetworkSchema);
-    return networks;
-  }
-};
+const NetworkModel = require("@models/Network");
 
 const validNetworks = async () => {
   const networks = await NetworkModel("airqo").distinct("name");
@@ -32,6 +21,20 @@ const validateNetwork = async (value) => {
   }
 };
 
+const validatePagination = (req, res, next) => {
+  // Retrieve the limit and skip values from the query parameters
+  const limit = parseInt(req.query.limit, 10);
+  const skip = parseInt(req.query.skip, 10);
+
+  // Validate and sanitize the limit value
+  req.query.limit = isNaN(limit) || limit < 1 ? 1000 : limit;
+
+  // Validate and sanitize the skip value
+  req.query.skip = isNaN(skip) || skip < 0 ? 0 : skip;
+
+  next();
+};
+
 const headers = (req, res, next) => {
   res.setHeader("Access-Control-Allow-Origin", "*");
   res.header(
@@ -42,6 +45,7 @@ const headers = (req, res, next) => {
   next();
 };
 router.use(headers);
+router.use(validatePagination);
 
 /******************* lessons *********************************************/
 router.get(
