+++ src/device-registry/routes/v2/measurements.js
@@ -5,19 +5,8 @@ const { check, oneOf, query, body, param } = require("express-validator");
 const constants = require("@config/constants");
 const mongoose = require("mongoose");
 const ObjectId = mongoose.Types.ObjectId;
-const { getModelByTenant } = require("@config/database");
 const { logElement, logText, logObject } = require("@utils/log");
-
-const NetworkSchema = require("@models/Network");
-const NetworkModel = (tenant) => {
-  try {
-    const networks = mongoose.model("networks");
-    return networks;
-  } catch (error) {
-    const networks = getModelByTenant(tenant, "network", NetworkSchema);
-    return networks;
-  }
-};
+const NetworkModel = require("@models/Network");
 
 const validNetworks = async () => {
   const networks = await NetworkModel("airqo").distinct("name");
@@ -31,6 +20,20 @@ const validateNetwork = async (value) => {
   }
 };
 
+const validatePagination = (req, res, next) => {
+  // Retrieve the limit and skip values from the query parameters
+  const limit = parseInt(req.query.limit, 10);
+  const skip = parseInt(req.query.skip, 10);
+
+  // Validate and sanitize the limit value
+  req.query.limit = isNaN(limit) || limit < 1 ? 1000 : limit;
+
+  // Validate and sanitize the skip value
+  req.query.skip = isNaN(skip) || skip < 0 ? 0 : skip;
+
+  next();
+};
+
 const headers = (req, res, next) => {
   res.setHeader("Access-Control-Allow-Origin", "*");
   res.header(
@@ -41,6 +44,7 @@ const headers = (req, res, next) => {
   next();
 };
 router.use(headers);
+router.use(validatePagination);
 
 /******************* create-event use-case *******************************/
 router.get(
@@ -53,7 +57,7 @@ router.get(
       .bail()
       .trim()
       .toLowerCase()
-      .isIn(["kcca", "airqo", "urban_better", "us_embassy", "nasa", "cross"])
+      .isIn(constants.NETWORKS)
       .withMessage("the tenant value is not among the expected ones"),
   ]),
   oneOf([
@@ -998,7 +1002,7 @@ router.post(
         .optional()
         .notEmpty()
         .toLowerCase()
-        .isIn(["kcca", "airqo", "urban_better", "usembassy", "nasa", "unep"])
+        .custom(validateNetwork)
         .withMessage("the network value is not among the expected ones"),
     ],
   ]),
